Split up single file, add code coverage capabilities
