+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 .idea/
 *.iml
+.nyc_output
+coverage
